Move environment var checker to SCons.Util
is_valid_construction_var was defined in Environment, but also used
in Variables. This meant a possibility of import loops, since Variables
used Environment for this function, and Environment uses Variables
to update from an object in Environment() and Clone().

According to breadcumbs in a benchmark script it used to be in Util
at some point in the past.
