+++ CHANGES.txt
@@ -86,8 +86,8 @@ RELEASE  VERSION/DATE TO BE FILLED IN LATER
       updated in 4.6). All SCons usage except unit test was already fully
       consistent with a bool.
     - When a variable is added to a Variables object, it can now be flagged
-      as "don't perform substitution" by setting the argument subst. 
-      This allows variables to contain characters which would otherwise 
+      as "don't perform substitution" by setting the argument subst.
+      This allows variables to contain characters which would otherwise
       cause expansion. Fixes #4241.
     - The test runner now recognizes the unittest module's return code of 5,
       which means no tests were run. SCons/Script/MainTests.py currently
@@ -95,6 +95,10 @@ RELEASE  VERSION/DATE TO BE FILLED IN LATER
       cause runtest to give up with an "unknown error code".
     - Updated the notes about reproducible builds with SCons and the example.
     - The Clone() method now respects the variables argument (fixes #3590)
+    - is_valid_construction_var (not past of the public API) moved from
+      Environment.py to Util to avoid the chance of import loops. Variables
+      and Environment both use the routine and Environment() uses a Variables()
+      object so better to move to a safter location.
 
 
 RELEASE 4.7.0 -  Sun, 17 Mar 2024 17:22:20 -0700
